+++ example/debouncer.dart
@@ -2,12 +2,12 @@ import 'dart:async';
 
 /// [Debouncer] is a stream transformer for debouncing pin events
 class Debouncer implements StreamTransformerBase<bool, bool> {
-  bool value;
-  bool _lastValue;
-  StreamController<bool> _controller;
+  bool? value;
+  bool? _lastValue;
+  StreamController<bool>? _controller;
 
   final Duration debounceDuration;
-  Timer _debounceTimer;
+  Timer? _debounceTimer;
 
   Debouncer(this.value, int debounceMilliseconds)
       : debounceDuration = Duration(milliseconds: debounceMilliseconds);
@@ -15,7 +15,7 @@ class Debouncer implements StreamTransformerBase<bool, bool> {
   @override
   Stream<bool> bind(Stream<bool> stream) {
     if (_controller != null) throw 'cannout use debouncer twice';
-    StreamSubscription<bool> subscription;
+    late StreamSubscription<bool> subscription;
     _controller = StreamController<bool>(onListen: () {
       subscription = stream.listen((bool newValue) {
         _lastValue = newValue;
@@ -23,14 +23,14 @@ class Debouncer implements StreamTransformerBase<bool, bool> {
         _debounceTimer = Timer(debounceDuration, () {
           if (value != _lastValue) {
             value = _lastValue;
-            _controller.add(_lastValue);
+            _controller!.add(_lastValue!);
           }
         });
       });
     }, onCancel: () {
       subscription.cancel();
     });
-    return _controller.stream;
+    return _controller!.stream;
   }
 
   StreamTransformer<RS, RT> cast<RS, RT>() =>