+++ test/debouncer_test.dart
@@ -8,22 +8,22 @@ const testPinNum = 2;
 
 void main() {
   var timeLimit = Duration(milliseconds: 10);
-  Completer<bool> completer;
-  Completer cancelCompleter;
+  Completer<bool>? completer;
+  late Completer cancelCompleter;
   var controller = StreamController<bool>(onCancel: () {
     cancelCompleter.complete();
   });
   var debouncer = Debouncer(false, 1);
   StreamSubscription subscription =
       controller.stream.transform(debouncer).listen((bool newValue) {
-    completer.complete(newValue);
+    completer!.complete(newValue);
     completer = null;
   });
 
   test('one event', () async {
     completer = Completer<bool>();
     controller.add(true);
-    var future = completer.future.timeout(timeLimit);
+    var future = completer!.future.timeout(timeLimit);
     expect(await future, true);
   });
 
@@ -32,7 +32,7 @@ void main() {
     controller.add(false);
     controller.add(true);
     try {
-      await completer.future.timeout(timeLimit);
+      await completer!.future.timeout(timeLimit);
       fail('expected timeout - should not be any events');
     } on TimeoutException {
       // Expected timeout... no events
@@ -45,7 +45,7 @@ void main() {
     controller.add(true);
     controller.add(false);
     // Expect one result ... no exceptions from additional events
-    expect(await completer.future.timeout(timeLimit), false);
+    expect(await completer!.future.timeout(timeLimit), false);
     await Future.delayed(timeLimit);
   });