+++ test/src/pwm_node_test.dart
@@ -16,27 +16,27 @@ void main() {
 
     addPwmNode(newNode);
     expect(firstPwmNode, same(newNode));
-    expect(firstPwmNode.nextNode, isNull);
-    expect(firstPwmNode.nextNode, isNull);
+    expect(firstPwmNode!.nextNode, isNull);
+    expect(firstPwmNode!.nextNode, isNull);
     expect(newNode.state, true);
     expect(currentPwmTick, 50);
 
     newNode.stateChanged = Completer<bool>();
-    expect(await newNode.stateChanged.future, isFalse);
+    expect(await newNode.stateChanged!.future, isFalse);
     expect(currentPwmTick, 0);
 
     newNode.stateChanged = Completer<bool>();
-    expect(await newNode.stateChanged.future, isTrue);
+    expect(await newNode.stateChanged!.future, isTrue);
     expect(currentPwmTick, 50);
 
     newNode.stateChanged = Completer<bool>();
-    expect(await newNode.stateChanged.future, isFalse);
+    expect(await newNode.stateChanged!.future, isFalse);
     expect(currentPwmTick, 0);
 
     var node = removePwmNode(8);
     expect(node, isNull);
     expect(firstPwmNode, same(newNode));
-    expect(firstPwmNode.nextNode, isNull);
+    expect(firstPwmNode!.nextNode, isNull);
 
     node = removePwmNode(7);
     expect(node, same(newNode));
@@ -53,44 +53,44 @@ void main() {
     addPwmNode(newNode2);
     addPwmNode(newNode3);
     expect(firstPwmNode, same(newNode2));
-    expect(firstPwmNode.nextNode, same(newNode1));
-    expect(firstPwmNode.nextNode.nextNode, same(newNode3));
-    expect(firstPwmNode.nextNode.nextNode.nextNode, isNull);
+    expect(firstPwmNode!.nextNode, same(newNode1));
+    expect(firstPwmNode!.nextNode!.nextNode, same(newNode3));
+    expect(firstPwmNode!.nextNode!.nextNode!.nextNode, isNull);
     expect(newNode1.state, false);
     expect(newNode2.state, false);
     expect(newNode3.state, false);
     expect(currentPwmTick, 0);
 
     newNode1.stateChanged = Completer<bool>();
-    expect(await newNode1.stateChanged.future, isTrue);
+    expect(await newNode1.stateChanged!.future, isTrue);
     expect(newNode1.state, true);
     expect(newNode2.state, true);
     expect(newNode3.state, true);
     expect(currentPwmTick, 25);
 
     newNode2.stateChanged = Completer<bool>();
-    expect(await newNode2.stateChanged.future, isFalse);
+    expect(await newNode2.stateChanged!.future, isFalse);
     expect(newNode1.state, true);
     expect(newNode2.state, false);
     expect(newNode3.state, true);
     expect(currentPwmTick, 50);
 
     newNode1.stateChanged = Completer<bool>();
-    expect(await newNode1.stateChanged.future, isFalse);
+    expect(await newNode1.stateChanged!.future, isFalse);
     expect(newNode1.state, false);
     expect(newNode2.state, false);
     expect(newNode3.state, true);
     expect(currentPwmTick, 75);
 
     newNode3.stateChanged = Completer<bool>();
-    expect(await newNode3.stateChanged.future, isFalse);
+    expect(await newNode3.stateChanged!.future, isFalse);
     expect(newNode1.state, false);
     expect(newNode2.state, false);
     expect(newNode3.state, false);
     expect(currentPwmTick, 0);
 
     newNode1.stateChanged = Completer<bool>();
-    expect(await newNode1.stateChanged.future, isTrue);
+    expect(await newNode1.stateChanged!.future, isTrue);
     expect(newNode1.state, true);
     expect(newNode2.state, true);
     expect(newNode3.state, true);
@@ -99,12 +99,12 @@ void main() {
     var node = removePwmNode(8);
     expect(node, isNull);
     expect(firstPwmNode, same(newNode2));
-    expect(firstPwmNode.nextNode, same(newNode1));
-    expect(firstPwmNode.nextNode.nextNode, same(newNode3));
-    expect(firstPwmNode.nextNode.nextNode.nextNode, isNull);
+    expect(firstPwmNode!.nextNode, same(newNode1));
+    expect(firstPwmNode!.nextNode!.nextNode, same(newNode3));
+    expect(firstPwmNode!.nextNode!.nextNode!.nextNode, isNull);
 
     newNode2.stateChanged = Completer<bool>();
-    expect(await newNode2.stateChanged.future, isFalse);
+    expect(await newNode2.stateChanged!.future, isFalse);
     expect(newNode1.state, true);
     expect(newNode2.state, false);
     expect(newNode3.state, true);
@@ -113,11 +113,11 @@ void main() {
     node = removePwmNode(7);
     expect(node, same(newNode1));
     expect(firstPwmNode, same(newNode2));
-    expect(firstPwmNode.nextNode, same(newNode3));
-    expect(firstPwmNode.nextNode.nextNode, isNull);
+    expect(firstPwmNode!.nextNode, same(newNode3));
+    expect(firstPwmNode!.nextNode!.nextNode, isNull);
 
     newNode3.stateChanged = Completer<bool>();
-    expect(await newNode3.stateChanged.future, isFalse);
+    expect(await newNode3.stateChanged!.future, isFalse);
     expect(newNode1.state, true); // removed before turned off
     expect(newNode2.state, false);
     expect(newNode3.state, false);
@@ -126,24 +126,24 @@ void main() {
     node = removePwmNode(11);
     expect(node, same(newNode3));
     expect(firstPwmNode, same(newNode2));
-    expect(firstPwmNode.nextNode, isNull);
+    expect(firstPwmNode!.nextNode, isNull);
 
     newNode2.stateChanged = Completer<bool>();
-    expect(await newNode2.stateChanged.future, isTrue);
+    expect(await newNode2.stateChanged!.future, isTrue);
     expect(newNode1.state, true); // removed before turned off
     expect(newNode2.state, true);
     expect(newNode3.state, false); // removed when off
     expect(currentPwmTick, 25);
 
     newNode2.stateChanged = Completer<bool>();
-    expect(await newNode2.stateChanged.future, isFalse);
+    expect(await newNode2.stateChanged!.future, isFalse);
     expect(newNode1.state, true); // removed before turned off
     expect(newNode2.state, false);
     expect(newNode3.state, false); // removed when off
     expect(currentPwmTick, 0);
 
     newNode2.stateChanged = Completer<bool>();
-    expect(await newNode2.stateChanged.future, isTrue);
+    expect(await newNode2.stateChanged!.future, isTrue);
     expect(newNode1.state, true); // removed before turned off
     expect(newNode2.state, true);
     expect(newNode3.state, false); // removed when off
@@ -151,13 +151,13 @@ void main() {
 
     addPwmNode(newNode3);
     expect(firstPwmNode, same(newNode2));
-    expect(firstPwmNode.nextNode, same(newNode3));
-    expect(firstPwmNode.nextNode.nextNode, isNull);
+    expect(firstPwmNode!.nextNode, same(newNode3));
+    expect(firstPwmNode!.nextNode!.nextNode, isNull);
 
     node = removePwmNode(9);
     expect(node, same(newNode2));
     expect(firstPwmNode, same(newNode3));
-    expect(firstPwmNode.nextNode, isNull);
+    expect(firstPwmNode!.nextNode, isNull);
   });
 
   tearDown(() {
@@ -167,7 +167,7 @@ void main() {
 
 class MockPwmNode extends PwmNode {
   bool state = false;
-  Completer<bool> stateChanged;
+  Completer<bool>? stateChanged;
 
   MockPwmNode(int bcmGpioPin, int dutyCycle) : super(bcmGpioPin, dutyCycle);